+++ source/commands/_commandDictionary.js
@@ -6,7 +6,7 @@ exports.commandSets = [
 	new CommandSet("General Commands", "These commands are general use utilities for the server.", false,
 		["at-channel.js", "timestamp.js", "roll.js", "petition.js"]),
 	new CommandSet("Informantional Commands", "Use these commands to learn more about this server or HorizonsBot.", false,
-		["rules.js", "commands.js", "roles.js", "list.js", "about.js", "version.js", "data-policy.js", "press-kit.js"]),
+		["info.js", "commands.js", "list.js", "version.js"]),
 	new CommandSet("Club Commands", "Clubs are private text and voice channels that include organization utilities like automatic reminders.", false,
 		["club-invite.js", "club-kick.js", "club-leave.js", "club-send-reminder.js", "club-config.js", "club-promote-host.js", "club-add.js", "club-sunset.js"]),
 	new CommandSet("Moderation Commands", "Commands for moderators. Required permissions are listed in (parenthesis) at the beginning of the description.", true,