+++ source/commands/post-reference.js
@@ -1,7 +1,7 @@
 const Command = require('../classes/Command.js');
 const { saveObject } = require('../helpers.js');
 const { referenceMessages: listMessages, buildListMessagePayload } = require('../engines/referenceEngine.js');
-const embed = require("../../config/embeds/rules.json");
+const { rulesEmbedBuilder, pressKitEmbedBuilder } = require('../engines/messageEngine.js');
 
 const customId = "post-reference";
 const options = [
@@ -9,7 +9,8 @@ const options = [
 		type: "String", name: "reference", description: "which message to post", required: true, choices: [
 			{ name: "the petiton list", value: "petition" },
 			{ name: "the club list", value: "club" },
-			{ name: "the rules embed", value: "rules" }
+			{ name: "the rules embed", value: "rules" },
+			{ name: "the press kit", value: "press-kit" }
 		]
 	}
 ];
@@ -28,7 +29,9 @@ module.exports.execute = async (interaction) => {
 			messageOptions = await buildListMessagePayload(interaction.guild.memberCount, listType);
 			break;
 		case "rules":
-			messageOptions = { embeds: [embed] };
+			messageOptions = { embeds: [rulesEmbedBuilder()] };
+		case "press-kit":
+			messageOptions = { embeds: [pressKitEmbedBuilder()] };
 			break;
 	}
 	interaction.channel.send(messageOptions).then(message => {