+++ source/engines/messageEngine.js
@@ -93,3 +93,55 @@ exports.versionEmbedBuilder = function () {
 		return embed.addFields({ name: "Other Discord Bots", value: "Check out other Imaginary Horizons Productions bots or commission your own on the [IHP GitHub](https://github.com/Imaginary-Horizons-Productions)" });
 	})
 }
+
+exports.rulesEmbedBuilder = function () {
+	return new EmbedBuilder().setColor(7045611)
+		.setAuthor({
+			"name": "Click here to visit HorizonsBot's GitHub",
+			"iconURL": "https://cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png",
+			"url": "https://github.com/Imaginary-Horizons-Productions/HorizonsBot"
+		})
+		.setTitle("Server Rules (/info rules)")
+		.setThumbnail("https://cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png")
+		.setDescription("Imaginary Horizons is an community that values dignity, creativity, and diversity. Here are our server's rules:")
+		.addFields({
+			"name": "Give the Benefit of the Doubt",
+			"value": "Always be prepared to see a conflict as a misunderstanding and don't intentionally pick fights. Whenever possible, try to hash out misunderstandings in private messages instead of in public chats."
+		},
+			{
+				"name": "No Ridicule/Verbal Abuse",
+				"value": "We value diversity of opinion and expression in this server. Attacking or making fun of people is not productive. This rule includes excessive self-deprecation, unreciprocated trash talk, as well as bigoted terms of any kind."
+			},
+			{
+				"name": "Start Discussions in the Appropriate Channel",
+				"value": "Please post in appropriate channels so we can keep discussion organized. If you can't find a good channel to post in, check the Channel and Roles Browser at the top of the channel list to see if you can opt-in. If you still can't find a good channel, you can /petition for one to be made."
+			},
+			{
+				"name": "Discord Usage Tips",
+				"value": "- Start your message with `@silent` to prevent it from waking people up in the middle of the night.\n- Surround text with `||` to mark it a spoiler. ||example||"
+			}
+		)
+		.setFooter(exports.randomEmbedFooter());
+}
+
+exports.pressKitEmbedBuilder = function () {
+	return new EmbedBuilder().setColor(7045611)
+		.setAuthor({
+			"name": "Click here to visit HorizonsBot's GitHub",
+			"iconURL": "https://cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png",
+			"url": "https://github.com/Imaginary-Horizons-Productions/HorizonsBot"
+		})
+		.setTitle("Imaginary Horizons Branding (/info press-kit)")
+		.setThumbnail("https://cdn.discordapp.com/icons/353575133157392385/c78041f52e8d6af98fb16b8eb55b849a.png")
+		.addFields({
+			"name": "Colors",
+			"value": "Salmon - #f07581\nPeriwinkle - #6b81eb\nSpring Green - #b0ffe8\nWhite - #ffffff"
+		},
+			{
+				"name": "Links",
+				"value": "Server Invite - discord.gg/5EPWvu4\ntop.gg vote page - https://top.gg/servers/353575133157392385/vote"
+			}
+		)
+		.setImage("https://cdn.discordapp.com/attachments/812099861084241982/1094738237169340558/Patreon_Banner_Final.jpg")
+		.setFooter(exports.randomEmbedFooter());
+}