+++ ChangeLog.md
@@ -1,6 +1,8 @@
 # HorizonsBot Change Log
 #### HorizonsBot Version 2.2.0:
-- Updated rules embed
+- Updated rules and data policy
+- Combined `/about`, `/data-policy`, `/press-kit`, `/roles`, & `/rules` into `/info`
+- Club events now use the club's image
 #### HorizonsBot Version 2.1.0:
 - Removed opt-in topics handling, please use the Channel & Roles Browser instead
    - Retained `/petition` for text channels